Added multithreading.
Added support for running unittest concurrently to aid with unit test
scaling. Currently support is rough and the test reporting is currently
not thread safe and therefore this cannot be used in its current state.
There are still a number of things that should be addressed in the
future.

TODO:
    * Specify number of threads at runtime (Requires cmdline support).
    * Make it possible to compile without thread support (currently
      this is mandatory)
    * Add better logging.
    * Modify scheduler to use dynamic scheduling.
    * Code cleanup. Currently the code is a mess.